recurring_tasks_service/requirements.txt
confluent-kafka
python-dotenv
sqlmodel
psycopg2-binary
python-dateutil
